refactored, added decorator and fixed auth state generation & storage
